requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
colorama==0.4.6
crispy-bootstrap5==0.7
dj-database-url==3.0.1
Django==4.2.7
django-cloudinary-storage==0.3.0
django-crispy-forms==2.1
django-extensions==3.2.3
djangorestframework==3.14.0
ecdsa==0.19.1
gunicorn==21.2.0
idna==3.10
MarkupSafe==3.0.2
packaging==25.0
Pillow==10.0.1
psycopg2-binary==2.9.10
python-barcode==0.16.1
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
qrcode==8.2
reportlab==4.4.4
requests==2.32.5
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.6.0
